Corrigindo a finalização do codigo

+++ src/main/java/com/generation/gamestore/controller/CategoriaController.java
@@ -1,20 +1,16 @@
 package com.generation.gamestore.controller;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 import com.generation.gamestore.model.Categoria;
 import com.generation.gamestore.repository.CategoriaRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import jakarta.validation.Valid;
+import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/categorias")
@@ -25,28 +21,43 @@ public class CategoriaController {
 	private CategoriaRepository categoriaRepository;
 
 	@GetMapping
-	public List<Categoria> listar() {
-		return categoriaRepository.findAll();
+	public ResponseEntity<List<Categoria>> getAll() {
+		return ResponseEntity.ok(categoriaRepository.findAll());
 	}
 
 	@GetMapping("/{id}")
-	public Categoria buscarPorId(@PathVariable Long id) {
-		return categoriaRepository.findById(id).orElseThrow();
+	public ResponseEntity<Categoria> getById(@PathVariable Long id) {
+		return categoriaRepository.findById(id).map(ResponseEntity::ok)
+				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+	}
+
+	@GetMapping("/nome/{nome}")
+	public ResponseEntity<List<Categoria>> getByNome(@PathVariable String nome) {
+		return ResponseEntity.ok(categoriaRepository.findAllByNomeContainingIgnoreCase(nome));
 	}
 
 	@PostMapping
-	public Categoria criar(@RequestBody Categoria categoria) {
-		return categoriaRepository.save(categoria);
+	public ResponseEntity<Categoria> post(@Valid @RequestBody Categoria categoria) {
+		return ResponseEntity.status(HttpStatus.CREATED).body(categoriaRepository.save(categoria));
 	}
 
-	@PutMapping("/{id}")
-	public Categoria atualizar(@PathVariable Long id, @RequestBody Categoria categoria) {
-		categoria.setId(id);
-		return categoriaRepository.save(categoria);
+	@PutMapping
+	public ResponseEntity<Categoria> put(@Valid @RequestBody Categoria categoria) {
+		if (categoriaRepository.existsById(categoria.getId())) {
+			return ResponseEntity.status(HttpStatus.OK).body(categoriaRepository.save(categoria));
+		}
+		throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Categoria não encontrada!");
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
-	public void deletar(@PathVariable Long id) {
+	public void delete(@PathVariable Long id) {
+		Optional<Categoria> categoria = categoriaRepository.findById(id);
+
+		if (categoria.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+
 		categoriaRepository.deleteById(id);
 	}
 }
