+++ src/main/java/com/generation/gamestore/controller/produtoController.java
@@ -1,20 +1,17 @@
 package com.generation.gamestore.controller;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 import com.generation.gamestore.model.Produto;
 import com.generation.gamestore.repository.ProdutoRepository;
+import com.generation.gamestore.repository.CategoriaRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import jakarta.validation.Valid;
+import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/produtos")
@@ -24,29 +21,53 @@ public class produtoController {
 	@Autowired
 	private ProdutoRepository produtoRepository;
 
+	@Autowired
+	private CategoriaRepository categoriaRepository;
+
 	@GetMapping
-	public List<Produto> listar() {
-		return produtoRepository.findAll();
+	public ResponseEntity<List<Produto>> getAll() {
+		return ResponseEntity.ok(produtoRepository.findAll());
 	}
 
 	@GetMapping("/{id}")
-	public Produto buscarPorId(@PathVariable Long id) {
-		return produtoRepository.findById(id).orElseThrow();
+	public ResponseEntity<Produto> getById(@PathVariable Long id) {
+		return produtoRepository.findById(id).map(ResponseEntity::ok)
+				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+	}
+
+	@GetMapping("/nome/{nome}")
+	public ResponseEntity<List<Produto>> getByNome(@PathVariable String nome) {
+		return ResponseEntity.ok(produtoRepository.findAllByNomeContainingIgnoreCase(nome));
 	}
 
 	@PostMapping
-	public Produto criar(@RequestBody Produto produto) {
-		return produtoRepository.save(produto);
+	public ResponseEntity<Produto> post(@Valid @RequestBody Produto produto) {
+		if (categoriaRepository.existsById(produto.getCategoria().getId())) {
+			return ResponseEntity.status(HttpStatus.CREATED).body(produtoRepository.save(produto));
+		}
+		throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Categoria não existe!");
 	}
 
-	@PutMapping("/{id}")
-	public Produto atualizar(@PathVariable Long id, @RequestBody Produto produto) {
-		produto.setId(id);
-		return produtoRepository.save(produto);
+	@PutMapping
+	public ResponseEntity<Produto> put(@Valid @RequestBody Produto produto) {
+		if (produtoRepository.existsById(produto.getId())) {
+			if (categoriaRepository.existsById(produto.getCategoria().getId())) {
+				return ResponseEntity.status(HttpStatus.OK).body(produtoRepository.save(produto));
+			}
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Categoria não existe!");
+		}
+		throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Produto não encontrado!");
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
-	public void deletar(@PathVariable Long id) {
+	public void delete(@PathVariable Long id) {
+		Optional<Produto> produto = produtoRepository.findById(id);
+
+		if (produto.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+
 		produtoRepository.deleteById(id);
 	}
 }
